requirements.txt
Flask==1.1.1
ifaddr==0.1.6
Pillow==7.0.0